Segment Ratings
Segments are used to apply ratings to the underlying ways. When a segment is uplaoded,
we save the segment metadata and then create a rating by doing a lookup of the ways
that are overlapped by the segment's polyline. These ratings are then saved. Rating summary
is a view that offers a single data entry per way to retrieve the averaged data about the way.