New Segment rating system
The old one was broken - it didn't allow for accurate ratings since it
tied the rating to the actual map road data. This new system has a
completely decoupled ratings from the map data. This allows for very
precise ratings. We do this through MiniSegments. See doc for more info.